feat(UI): cập nhật thành nhắn tin với group cập nhật cấu trúc giao diện, gộp giao diện nhắn tin của sinh viên và giảng viên làm một
