+++ public/demo.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+</head>
+<body class="box">
+    <div style="height: 1000px;"></div>
+</body>
+</html>
+<style>
+    .box {
+        height: 100%;
+        background-image: linear-gradient(45deg,  #ffffff, #ff0000);
+        background-size: cover;
+        background-repeat: no-repeat;
+    }
+</style>
